implemented selling items in shop
-fixed a bug where shopInventory wouldn't display items proparly;
-fixed inventory animation;
-added selling items;

+++ Assets/Scripts/Player/Item.cs
@@ -8,7 +8,7 @@ public class Item : ScriptableObject
     public string itemName = "New Item";
     public string itemDescription = "New Description";
     public Sprite icon;
-    public int price = 0;
+    public float price = 0f;
     public enum Type { Default, Consumable, Weapon, Ammunition }
     public Type type = Type.Default;
 }
